% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{create_example_gene_sets}
\alias{create_example_gene_sets}
\title{Create Simple Example Gene Sets}
\usage{
create_example_gene_sets(gene_universe, n_pathways = 50, pathway_size = 30)
}
\arguments{
\item{gene_universe}{Vector of all available gene names}

\item{n_pathways}{Number of pathways to create}

\item{pathway_size}{Average size of each pathway}
}
\value{
Named list of example gene sets
}
\description{
Creates example gene sets for testing purposes.
}
\examples{
data(eset)
# Create example gene sets
example_sets <- create_example_gene_sets(rownames(eset), n_pathways = 20)
}
