% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset.R
\docType{data}
\name{eset}
\alias{eset}
\title{Example Expression Set}
\format{
A matrix with 10 rows and 10 columns.
}
\usage{
data(eset)
}
\description{
A toy dataset for demonstrating clustering.
}
\examples{
data(eset)
str(eset)
}
\keyword{datasets}
