% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{perform_normalization}
\alias{perform_normalization}
\title{Perform normalization on an expression matrix}
\usage{
perform_normalization(expr, method)
}
\arguments{
\item{expr}{Numeric matrix of expression values (genes x samples).}

\item{method}{Normalization method, one of "log2", "quantile", "zscore", or "deseq".}
}
\value{
Normalized expression matrix.
}
\description{
Supports "log2", "quantile", "zscore", and "deseq" normalization methods.
}
